frontend/requirements.txt
Flask
pycryptodomex
argon2-cffi
pyotp
requests
Pillow
qrcode[pil]
